now make reveal query connection an option to the command center
