+++ typings/index.ts
@@ -163,6 +163,7 @@ export module SqluiEnums {
     | 'clientEvent/query/close'
     | 'clientEvent/query/closeCurrentlySelected'
     | 'clientEvent/query/closeOther'
+    | 'clientEvent/query/reveal'
     | 'clientEvent/session/new'
     | 'clientEvent/session/rename'
     | 'clientEvent/session/switch';
